Melhora forma como o game faz requisicoes rest

+++ Assets/Scripts/Entity/User.cs
@@ -3,5 +3,5 @@
 [Serializable]
 public class User {
 	public string nickname;
-	public float money;
+	public long money;
 }
